pyproject.toml
[tool.poetry]
name = "flower-shop-users"
version = "0.1.0"
description = ""
authors = ["Dzmitry Bakunovich <dv.bakunovich@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.4"
djangorestframework = "^3.13.1"
djangorestframework-simplejwt = "^5.1.0"
psycopg2-binary = "^2.9.3"
Pillow = "^9.1.0"
celery = "^5.2.6"
redis = "^4.2.2"
ipython = "^8.2.0"
drf-spectacular = "^0.22.0"
dj-database-url = "^0.5.0"
django-cors-headers = "^3.11.0"
djangorestframework-camel-case = "^1.3.0"
pytest-django = "^4.5.2"
factory-boy = "^3.2.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
